Add a route for adding a product to the store
Moved the database connect class to its own file because
it is enbles to import it to other controller files like inventory.
Specified a route that calls an inventory member function and created
a function for app.py to see if an element is part of any array of
records or a single record passed.